fix(manage-products): fix bug for create product and grant permission when delete, update product.

+++ GoodsExchange.BusinessLogic/Services/ProductService.cs
@@ -1,20 +1,20 @@
 using GoodsExchange.BusinessLogic.Common;
+using GoodsExchange.BusinessLogic.Extensions;
 using GoodsExchange.BusinessLogic.RequestModels.Product;
-using GoodsExchange.BusinessLogic.ViewModels;
 using GoodsExchange.BusinessLogic.ViewModels.Product;
 using GoodsExchange.Data.Context;
 using GoodsExchange.Data.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 
 namespace GoodsExchange.BusinessLogic.Services
 {
-
     public interface IProductService
     {
         Task<ApiResult<ProductViewModel>> CreateProduct(CreateProductRequestModel request);
         Task<ApiResult<ProductViewModel>> UpdateProduct(UpdateProductRequestModel request);
         Task<ApiResult<bool>> DeleteProduct(Guid id);
-        Task<PageResult<ProductViewModel>> GetProductsAsync(PagingRequestModel request, SearchRequestModel search, GetAllProductRequestModel model);
+        Task<PageResult<ProductViewModel>> GetProductsAsync(PagingRequestModel request, SearchRequestModel search, GetAllProductRequestModel model, bool seller = false);
         Task<ApiResult<ProductDetailsViewModel>> GetById(Guid id);
         Task<ApiResult<bool>> UpdateProductStatus(Guid id, bool status);
         Task<ApiResult<bool>> ApproveProduct(Guid id);
@@ -24,10 +24,12 @@ namespace GoodsExchange.BusinessLogic.Services
     {
         private readonly GoodsExchangeDbContext _context;
         private readonly IRatingService _ratingService;
-        public ProductService(GoodsExchangeDbContext context, IRatingService ratingService)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public ProductService(GoodsExchangeDbContext context, IRatingService ratingService, IHttpContextAccessor httpContextAccessor)
         {
             _context = context;
             _ratingService = ratingService;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public async Task<ApiResult<bool>> ApproveProduct(Guid id)
@@ -35,38 +37,58 @@ namespace GoodsExchange.BusinessLogic.Services
             var product = await _context.Products.FindAsync(id);
             if (product == null)
             {
-                return new ApiErrorResult<bool>("Product does not exists.");
+                return new ApiErrorResult<bool>("Product does not exist.");
             }
             product.IsApproved = true;
             product.ApprovedDate = DateTime.Now;
             await _context.SaveChangesAsync();
 
-            return new ApiSuccessResult<bool>();
+            return new ApiSuccessResult<bool>(true);
         }
 
         public async Task<ApiResult<ProductViewModel>> CreateProduct(CreateProductRequestModel request)
         {
+            var existingProduct = await _context.Products.FirstOrDefaultAsync(p => p.ProductName == request.ProductName);
+            if (existingProduct != null)
+            {
+                return new ApiErrorResult<ProductViewModel>("Product name already exists.");
+            }
+
+            var category = await _context.Categories.FirstOrDefaultAsync(c => c.CategoryId == request.CategoryId);
+            if (category == null)
+            {
+                return new ApiErrorResult<ProductViewModel>("Category does not exist.");
+            }
+
             var product = new Product()
             {
                 ProductName = request.ProductName,
                 Description = request.Description,
                 ProductImageUrl = request.ProductImageUrl,
                 Price = request.Price,
-                CategoryId = request.CategoryId,
+                IsActive = true,
                 UploadDate = DateTime.Now,
-                IsActive = false
+                UserUploadId = Guid.Parse(_httpContextAccessor.GetCurrentUserId()),
+                IsApproved = false,
+                CategoryId = category.CategoryId
             };
             await _context.Products.AddAsync(product);
             await _context.SaveChangesAsync();
+
+            var user = await _context.Users.FirstOrDefaultAsync(u => u.UserId == product.UserUploadId);
+
             var result = new ProductViewModel()
             {
+                ProductId = product.ProductId,
                 ProductName = product.ProductName,
                 Description = product.Description,
                 ProductImageUrl = product.ProductImageUrl,
                 Price = product.Price,
-                CategoryName = product.Category.CategoryName,
+                IsActive = product.IsActive,
+                UserUpload = user.UserName,
                 UploadDate = product.UploadDate,
-                Status = product.IsActive
+                ApprovedDate = product.ApprovedDate,
+                CategoryName = _context.Categories.FirstOrDefault(c => c.CategoryId == product.CategoryId).CategoryName,
             };
 
             return new ApiSuccessResult<ProductViewModel>(result);
@@ -77,16 +99,21 @@ namespace GoodsExchange.BusinessLogic.Services
             var product = await _context.Products.FindAsync(id);
             if (product == null)
             {
-                return new ApiErrorResult<bool>("Product does not exists.");
+                return new ApiErrorResult<bool>("Product does not exist.");
+            }
+
+            var userId = Guid.Parse(_httpContextAccessor.GetCurrentUserId());
+            if (product.UserUploadId != userId)
+            {
+                return new ApiErrorResult<bool>("You do not have permission to delete this product.");
             }
 
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
 
-            return new ApiSuccessResult<bool>();
+            return new ApiSuccessResult<bool>(true);
         }
-
-        public async Task<PageResult<ProductViewModel>> GetProductsAsync(PagingRequestModel paging, SearchRequestModel search, GetAllProductRequestModel model)
+        public async Task<PageResult<ProductViewModel>> GetProductsAsync(PagingRequestModel paging, SearchRequestModel search, GetAllProductRequestModel model, bool seller = false)
         {
             var query = _context.Products.Where(p => p.UserUpload.IsActive == true).AsQueryable();
 
@@ -111,10 +138,6 @@ namespace GoodsExchange.BusinessLogic.Services
             {
                 query = query.Where(p => p.Price <= model.MaxPrice);
             }
-            if (model.Status != null)
-            {
-                query = query.Where(p => p.IsActive == model.Status);
-            }
             if (model.StartUploadDate != null)
             {
                 query = query.Where(p => p.UploadDate >= model.StartUploadDate);
@@ -122,10 +145,6 @@ namespace GoodsExchange.BusinessLogic.Services
             if (model.EndUploadDate != null)
             {
                 query = query.Where(p => p.UploadDate <= model.EndUploadDate);
-            }
-            if (model.IsApproved != null)
-            {
-                query = query.Where(p => p.IsApproved == model.IsApproved);
             }
             if (model.StartApprovedDate != null)
             {
@@ -168,22 +187,28 @@ namespace GoodsExchange.BusinessLogic.Services
 
             #endregion
 
+            if (seller == true)
+            {
+                query = query.Where(p => p.UserUploadId == Guid.Parse(_httpContextAccessor.GetCurrentUserId()));
+            }
+
             var totalItems = await query.CountAsync();
             var totalPages = (int)Math.Ceiling((double)totalItems / paging.PageSize);
 
             var data = await query.Skip((paging.PageIndex - 1) * paging.PageSize)
                                 .Take(paging.PageSize)
-                                .Select(p => new ProductViewModel()
+                                .Select(product => new ProductViewModel()
                                 {
-                                    ProductId = p.ProductId,
-                                    ProductName = p.ProductName,
-                                    Description = p.Description,
-                                    ProductImageUrl = p.ProductImageUrl,
-                                    Status = p.IsActive,
-                                    Price = p.Price,
-                                    CategoryName = p.Category.CategoryName,
-                                    ApprovedDate = p.ApprovedDate,
-                                    UploadDate = p.UploadDate,
+                                    ProductId = product.ProductId,
+                                    ProductName = product.ProductName,
+                                    Description = product.Description,
+                                    ProductImageUrl = product.ProductImageUrl,
+                                    Price = product.Price,
+                                    IsActive = product.IsActive,
+                                    UserUpload = _context.Users.FirstOrDefault(u => u.UserId == product.UserUploadId).UserName,
+                                    UploadDate = product.UploadDate,
+                                    ApprovedDate = product.ApprovedDate,
+                                    CategoryName = _context.Categories.FirstOrDefault(c => c.CategoryId == product.CategoryId).CategoryName
                                 })
                                 .ToListAsync();
 
@@ -196,32 +221,47 @@ namespace GoodsExchange.BusinessLogic.Services
             return result;
         }
 
-
         public async Task<ApiResult<ProductViewModel>> UpdateProduct(UpdateProductRequestModel request)
         {
             var product = await _context.Products.FindAsync(request.ProductId);
             if (product == null)
             {
-                return new ApiErrorResult<ProductViewModel>("Product does not exists.");
+                return new ApiErrorResult<ProductViewModel>("Product does not exist.");
             }
+
+            var userId = Guid.Parse(_httpContextAccessor.GetCurrentUserId());
+            if (product.UserUploadId != userId)
+            {
+                return new ApiErrorResult<ProductViewModel>("You do not have permission to update this product.");
+            }
+
+            var existingProduct = await _context.Products.FirstOrDefaultAsync(p => p.ProductName == request.ProductName);
+            if (existingProduct != null)
+            {
+                return new ApiErrorResult<ProductViewModel>("Product name already exists.");
+            }
+
 
             product.ProductName = request.ProductName;
             product.Description = request.Description;
             product.ProductImageUrl = request.ProductImageUrl;
             product.Price = request.Price.Value;
-            product.IsActive = request.Status.Value;
-            product.CategoryId = request.CategoryId;
+            product.CategoryId = request.CategoryId.Value;
 
             await _context.SaveChangesAsync();
+
             var result = new ProductViewModel()
             {
+                ProductId = product.ProductId,
                 ProductName = product.ProductName,
                 Description = product.Description,
                 ProductImageUrl = product.ProductImageUrl,
                 Price = product.Price,
-                CategoryName = product.Category.CategoryName,
+                IsActive = product.IsActive,
+                UserUpload = _context.Users.FirstOrDefault(u => u.UserId == product.UserUploadId).UserName,
                 UploadDate = product.UploadDate,
-                Status = product.IsActive
+                ApprovedDate = product.ApprovedDate,
+                CategoryName = _context.Categories.FirstOrDefault(c => c.CategoryId == product.CategoryId).CategoryName
             };
 
             return new ApiSuccessResult<ProductViewModel>(result);
@@ -232,13 +272,13 @@ namespace GoodsExchange.BusinessLogic.Services
             var product = await _context.Products.FindAsync(id);
             if (product == null)
             {
-                return new ApiErrorResult<bool>("Product does not exists.");
+                return new ApiErrorResult<bool>("Product does not exist.");
             }
 
             product.IsActive = status;
             await _context.SaveChangesAsync();
 
-            return new ApiSuccessResult<bool>();
+            return new ApiSuccessResult<bool>(true);
         }
 
         public async Task<ApiResult<ProductDetailsViewModel>> GetById(Guid id)
@@ -246,8 +286,10 @@ namespace GoodsExchange.BusinessLogic.Services
             var product = await _context.Products.FindAsync(id);
             if (product == null)
             {
-                return new ApiErrorResult<ProductDetailsViewModel>("Product does not exists.");
+                return new ApiErrorResult<ProductDetailsViewModel>("Product does not exist.");
             }
+
+            var user = await _context.Users.FirstOrDefaultAsync(u => u.UserId == product.UserUploadId);
 
             var result = new ProductDetailsViewModel()
             {
@@ -257,11 +299,11 @@ namespace GoodsExchange.BusinessLogic.Services
                 ProductImageUrl = product.ProductImageUrl,
                 ApprovedDate = product.ApprovedDate,
 
-                UserUpload = product.UserUpload.UserName,
-                UserImageUrl = product.UserUpload.UserImageUrl,
+                UserUpload = user.UserName,
+                UserImageUrl = user.UserImageUrl,
                 NumberOfRatings = _ratingService.CountNumberRatingOfUser(product.UserUpload.UserId).Result,
                 AverageNumberStars = _ratingService.CountAverageNumberStarsOfUser(product.UserUpload.UserId).Result,
-                UserPhoneNumber = product.UserUpload.PhoneNumber
+                UserPhoneNumber = user.PhoneNumber
             };
 
             return new ApiSuccessResult<ProductDetailsViewModel>(result);
